requirements.txt
discord.py==1.4.1
Django==3.1.1
pymongo==3.11.0
python-dotenv==0.14.0
requests==2.24.0
dnspython
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
pytz==2020.1
sqlparse==0.4.0
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.6.0
